sessions

+++ app/Controllers/Home.php
@@ -24,10 +24,17 @@ class Home extends BaseController
     public function pessoas(){
         $model = new PessoasModel();
 
+
+
         $data = [
             'title'=>'Pessoas',
-            'pessoas'=>$model->getPessoas()
+            'pessoas'=>$model->getPessoas(),
+            'session' =>\Config\Services::session()
         ];
+
+        if(!$data['session']->get('logado')){
+            return redirect("login");
+        }
 
         echo view('template/header');
         echo view('pessoa',$data);
@@ -35,12 +42,26 @@ class Home extends BaseController
     }
 
     public function cadastro(){
+
+        $data['session'] = \Config\Services::session();
+
+        if(!$data['session']->get('logado')){
+            return redirect("login");
+        }
+
         echo view('template/header');
         echo view('cadastro-pessoas');
         echo view('template/footer');
     }
 
     public function gravar(){
+
+        $data['session'] = \Config\Services::session();
+
+        if(!$data['session']->get('logado')){
+            return redirect("login");
+        }
+
         $model = new PessoasModel();
 
         $model->save([
@@ -54,6 +75,13 @@ class Home extends BaseController
     }
 
     public function excluir($id = null){
+
+        $data['session'] = \Config\Services::session();
+
+        if(!$data['session']->get('logado')){
+            return redirect("login");
+        }
+
         $model = new PessoasModel();
         $model->delete($id);
         return redirect("pessoa");
@@ -63,8 +91,13 @@ class Home extends BaseController
         $model = new PessoasModel();
 
         $data = [
-            'pessoa' => $model->getPessoa($id)
+            'pessoa' => $model->getPessoa($id),
+            'session' => \Config\Services::session()
         ];
+
+        if(!$data['session']->get('logado')){
+            return redirect("login");
+        }
 
         echo view('template/header');
         echo view('cadastro-pessoas',$data);
@@ -84,11 +117,23 @@ class Home extends BaseController
          $nome = $this->request->getVar("nome");
 
          $data['usario'] = $model->userLogin($nome, $senha);
+         $data['session'] = \Config\Services::session();
 
          if(empty($data['usario'])){
              return redirect("login");
          }else{
+             $sessionData = [
+                 'usuario' => $nome,
+                 'logado' => TRUE
+             ];
+             $data['session']->set($sessionData);
              return redirect("pessoa");
          }
+    }
+
+    public function sair(){
+        $data['session'] = \Config\Services::session();
+        $data['session']->destroy();
+        return redirect("login");
     }
 }
